standalone/experiments/multi-agent/distribute-multi-agent-system/requirement.txt
strands-agents>=0.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
botocore==1.34.0
pydantic==2.5.0
python-dotenv==1.0.0
asyncio-throttle==1.0.2
aiofiles==23.2.1
typing-extensions==4.8.0